requirements.txt
amqp==5.1.1
asgiref==3.3.4
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.21.27
botocore==1.24.27
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.4
django-admin-logs==1.0.2
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==21.1
django-heroku==0.3.1
django-image-optimizer==1.0.3
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-notifications-rest==0.0.1
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
jsonfield==3.1.0
kombu==5.2.4

packaging==21.3
Pillow
prompt-toolkit==3.0.31
# psycopg2==2.9.1
pyparsing==3.0.9
python-barcode==0.13.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-resize-image==1.1.20
pytz==2022.5
redis==4.3.4
requests==2.26.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.1
swapper==1.2.0
tinify==1.6.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.11.2
selenium
python-decouple

# ==============================================for anzara
# mysqlclient

# ===============================================for Azran / ELOR
